% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{Betaestglm.func}
\alias{Betaestglm.func}
\title{Beta Estimation for Initial Values}
\usage{
Betaestglm.func(
  Y,
  R,
  Gamma,
  A,
  Theta,
  X,
  Zmat,
  indices_R,
  lengths_R,
  Tp,
  K,
  px,
  pz,
  N,
  J,
  par,
  n.cores,
  ext,
  gamma_fix
)
}
\value{
A matrix containing the estimated Beta coefficients for each event.
}
\description{
This function estimates the initial values for Beta (and V) using Generalized Linear Models (GLMs).
It supports parallel processing and is designed for internal package use.
}
\seealso{
\code{lvhml_est_fit_K} for the description of parameters.
}
\keyword{internal}
