% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{Betaonlyparglm.func}
\alias{Betaonlyparglm.func}
\title{Function to run glm function to get estimate of beta.}
\usage{
Betaonlyparglm.func(
  yris,
  indices_R,
  lengths_R,
  X,
  N,
  px,
  K,
  Tp,
  ext,
  gamma_fix
)
}
\arguments{
\item{yris}{(Tp x N) Matrix containing outcomes from Y associated with the jth event.
See \code{\link{lvhml_est}} for the meaning and dimensions of remaining arguments.}
}
\description{
Function to run glm function to get estimate of beta.
}
