% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_var.R
\name{calasympvar.func}
\alias{calasympvar.func}
\title{Compute Asymptotic Variance of uj for all j.}
\usage{
calasympvar.func(X, Z, Thetahat, Uhat, R, J, N, Tp)
}
\value{
An array containing the estimated asymptotic variance for each event.
}
\description{
Calculates the asymptotic variance of the U parameter for each event (j).
}
\keyword{#'@seealso}
\keyword{\code{lvhml_est_fit_K}}
\keyword{description}
\keyword{for}
\keyword{internal}
\keyword{of}
\keyword{parameters.}
\keyword{the}
