% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{calCI_thi.func}
\alias{calCI_thi.func}
\title{Function to compute confidence interval for theta_i based on paramters estimate and data:}
\usage{
calCI_thi.func(X, Z, Thetahat, Uhat, Ahat, R, J, N, Tp, ext, gamma_fix)
}
\arguments{
\item{X, Z, Thetahat, Uhat, Ahat, R, J, N, Tp, ext, gamma_fix}{See \code{\link{lvhml_est}} for the meaning and dimensions of
these arguments.}
}
\description{
Function to compute confidence interval for theta_i based on paramters estimate and data:
}
