% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{calCI_uj.func}
\alias{calCI_uj.func}
\title{Function to compute confidence interval based on parameters estimate and data:}
\usage{
calCI_uj.func(X, Z, Thetahat, Uhat, R, J, N, Tp, ext, gamma_fix)
}
\arguments{
\item{X, Z, Thetahat, Uhat, R, J, N, Tp, ext, gamma_fix}{See \code{\link{lvhml_est}}.}
}
\description{
Function to compute confidence interval based on parameters estimate and data:
}
