% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_var.R
\name{calEEstore.func}
\alias{calEEstore.func}
\title{Function to calculate e_{ir}e_{ir}^{\top} for all combinations of i and r.}
\usage{
calEEstore.func(E, ujlen, N, Tp)
}
\arguments{
\item{E}{Numeric array of dimension (N x ujlen x Tp) containing time-adjusted intercepts, covariates and factors.}

\item{ujlen}{Length of \code{u_j}.}
}
\description{
Function to calculate e_{ir}e_{ir}^{\top} for all combinations of i and r.
}
