% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_var.R
\name{calPhi_j.func}
\alias{calPhi_j.func}
\title{Compute Phi_j}
\usage{
calPhi_j.func(EEstarstore, ujE, R, N, J, Tp, ujlen)
}
\arguments{
\item{EEstarstore}{A 4 dimensional array (ujlen x ujlen x N x Tp) storing pre-computed values of e_{ir}e_{ir}^{t} for all i and r.}

\item{ujE}{(N x Tp) matrix of uj times e_{ir} values for a specific j.}

\item{ujlen}{Length of the uj.}
}
\value{
The computed Phi_j matrix.
}
\description{
Computes the Phi_j matrix for the calculation of asymptotic variance, based on EEstarstore and ujE values.
}
\seealso{
\code{lvhml_est_fit_K} for the description of other parameters.

#' @keywords internal
}
