% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{calPsi_i.func}
\alias{calPsi_i.func}
\title{Compute Psi_i matrix for a single subject}
\usage{
calPsi_i.func(AAhatstore, Uthihat, ri, N, J, Tp, K)
}
\arguments{
\item{AAhatstore}{Numeric array of dimension (K x K x J x Tp) storing transformed loading products.}

\item{ri}{Vector of length Tp containing the missing indicator of the ith subject.
See \code{\link{lvhml_est}} for the meaning and dimensions of remaining arguments.}
}
\value{
The computed Psi_i matrix.
}
\description{
Compute Psi_i matrix for a single subject
}
