% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_var.R
\name{compute_Ehat}
\alias{compute_Ehat}
\title{Function to compute Ehat}
\usage{
compute_Ehat(Thetahat, X, Z, ujlen, K, Tp, px, pz, N, ext, gamma_fix)
}
\arguments{
\item{Thetahat, X, Z, ujlen, K, Tp, px, pz, N, ext, gamma_fix}{See \code{\link{lvhml_est_fit_K}}.}
}
\description{
Function to compute Ehat
}
