% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{compute_M}
\alias{compute_M}
\title{Helper: compute intermediate M-matrix from reconstructed data}
\usage{
compute_M(tildelijts, epsilon)
}
\description{
Helper: compute intermediate M-matrix from reconstructed data
}
