% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{estglmbeta.func}
\alias{estglmbeta.func}
\title{Functions to run glm beta. Serve comparison in simulation only. Corresponds to LR in manuscript.}
\usage{
estglmbeta.func(
  Y,
  R,
  X = matrix(0, nrow = 0, ncol = 0),
  Tp,
  par = 0,
  n.cores = 1,
  ext,
  gamma_fix
)
}
\arguments{
\item{Y, }{R, X , Tp, par, n.cores, ext, gamma_fix S
See \code{\link{lvhml_est}} for the meaning and dimensions of these arguments.}
}
\description{
Functions to run glm beta. Serve comparison in simulation only. Corresponds to LR in manuscript.
}
