% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{estglmm.func}
\alias{estglmm.func}
\title{Functions to run glmm(Generalised linear mixed models. In particular, logistic regression with random intercepts.) Corresponds to LRRI in manuscript
#' See \code{\link{lvhml_est}} for the meaning and dimensions of arguments.}
\usage{
estglmm.func(
  Y,
  R,
  X = matrix(0, nrow = 0, ncol = 0),
  Tp,
  par = 0,
  n.cores = 1,
  ext,
  gamma_fix
)
}
\description{
Functions to run glmm(Generalised linear mixed models. In particular, logistic regression with random intercepts.) Corresponds to LRRI in manuscript
#' See \code{\link{lvhml_est}} for the meaning and dimensions of arguments.
}
