% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{eval.func}
\alias{eval.func}
\title{Functions to evaluate performance of the estimator.}
\usage{
eval.func(
  Thetahat,
  Uhat,
  Thetastar,
  Ustar,
  X,
  Z,
  N,
  J,
  Tp,
  px,
  pz,
  K,
  ext,
  gamma_fix
)
}
\arguments{
\item{Thetahat, Uhat, Thetastar, Ustar, X, Z, N, J, Tp, px, pz, K, ext, }{gamma_fix
 See \code{\link{lvhml_est}} and
\code{\link{lvhml_ex_data}} for the meaning and dimensions of
these quantities.}
}
\description{
Functions to evaluate performance of the estimator.
}
