% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_example_data.R
\name{gendata.func}
\alias{gendata.func}
\title{Generate observed data given model probabilities}
\usage{
gendata.func(Lambda, Kstar, N, J, Tp)
}
\arguments{
\item{Lambda}{Numeric array of dimension N x J x Tp
containing success probabilities for all subjects, events and time points.}

\item{Kstar}{True number of latent factors}

\item{N}{Number of subjects.}

\item{J}{Number of events/items.}

\item{Tp}{Number of time points.}
}
\value{
A list with two components:
  \itemize{
    \item Y: N x J x Tp  array of binary responses NA for unobserved entries.
    \item R: N x Tp matrix of missingness indicators (1 for observed, 0 for missing).
  }
}
\description{
Internal helper used by \code{\link{lvhml_ex_data}}
  to simulate binary responses and missingness indicators from the
  model probabilities.
}
\keyword{internal}
