% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{GetZmat.func}
\alias{GetZmat.func}
\title{Function to get Zmat}
\usage{
GetZmat.func(Z, pz, N, Tp)
}
\arguments{
\item{Z}{(N x pz x Tp) Array of time-dependent covariates.}
}
\description{
Function to get Zmat
}
\seealso{
\code{lvhml_est_fit_K} for the description of other parameters.
}
\keyword{internal}
