% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{lvhdr_est}
\alias{lvhdr_est}
\title{Estimation Function for LVHDR Model}
\usage{
lvhdr_est(
  Y,
  R,
  X,
  Kset = 1:10,
  par = FALSE,
  n.cores = 1,
  Asymp = FALSE,
  Silent = FALSE,
  Z = array(0, dim = c(0, 0, 0))
)
}
\arguments{
\item{Y}{A N x J x Tp array storing occurrences of events (1 for occurred, 0 for no occurrence). Entries for unobserved data should be marked as NA.}

\item{R}{A N x Tp matrix representing the missing indicator (1 for observed, 0 for missing).}

\item{X}{A N x px matrix storing static covariates.}

\item{Kset}{A vector containing the candidate numbers of factors. The default value is 1:10.}

\item{par}{Logical. Indicates whether to use parallel computing. Default is FALSE.}

\item{n.cores}{The number of cores to use for parallel computing. Used in conjunction with par = TRUE. Default is 1.}

\item{Asymp}{Logical. Indicates whether to estimate the asymptotic variance. Default is FALSE.}

\item{Silent}{Logical. Option to suppress output showing progress. Default is FALSE.}

\item{Z}{A N x J x Tp array for time-dependent covariates. Defaults as empty array if not provided.}
}
\value{
A list containing the results of the estimation, including:
        \item{Thetahat}{Estimated Theta}
        \item{Gammahat}{Estimated Gamma}
        \item{Betahat}{Estimated Beta}
        \item{Ahat}{Estimated A}
        \item{Khat}{Estimated K}
        \item{ICs}{Information criteria for values of K in Kset}
        \item{Sigma_E}{Estimated asymptotic variance matrices for sqrt(N)*betahat_j (if Asymp = TRUE)}
        \item{SE}{The standard errors of Betahat (if Asymp = TRUE)}
}
\description{
Main function to implement the proposed estimator. This includes choosing the number of factors (K) and calculating asymptotic variances if required. The function is capable of handling time-dependent covariates.
}
\examples{
# Generate example dataset
dataset <- lvhdr_ex_data(J = 100, nset = "s", Kstar = 3)
attach(dataset)
out <- lvhdr_est(Y, R, X, Kset = 1:10, par = FALSE, n.cores = 1, Asymp = FALSE, Silent = FALSE, Z = array(0, dim = c(0, 0, 0)))

}
