% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_par.R
\name{lvhdr_est_fit_K}
\alias{lvhdr_est_fit_K}
\title{Internal Estimation Function for LVHDR Model}
\usage{
lvhdr_est_fit_K(
  Y,
  Ymatlog,
  R,
  XZ,
  X,
  XZmat,
  indices_R,
  lengths_R,
  Tp,
  K,
  cov_len,
  px,
  N,
  J,
  Tinit,
  Uinit,
  par,
  n.cores,
  Silent
)
}
\arguments{
\item{Ymatlog}{Logical matrix derived from Y to reduce memory usage.}

\item{XZ}{Combined static and time-dependent covariates.}

\item{XZmat}{Matrix form of combined X and Z covariates for internal use.}

\item{indices_R}{Indices of observed entries in R.}

\item{lengths_R}{Number of observed time points for each subject.}

\item{Tp}{Number of time points.}

\item{K}{Number of factors to estimate.}

\item{cov_len}{Length of combined covariates.}

\item{px}{Number of static covariates.}

\item{N}{Number of subjects.}

\item{J}{Number of events.}

\item{Tinit}{Initial values for Theta.}

\item{Uinit}{Initial values for U (Gamma, A, Beta).}
}
\value{
A list containing the estimation results, including the success status, estimates of Theta and U, and the number of iterations.
}
\description{
This function estimates A, Beta, and Theta for a given value of K. It's designed
to be used internally within the package and supports parallel computing, handling
of time-dependent covariates, and convergence checking.
}
\seealso{
\code{\link{lvhdr_est}} for the description of \code{Y}, \code{R}, \code{X}, \code{par}, \code{n.cores} and \code{Silent}.
}
\keyword{internal}
