% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{lvhdr_norm}
\alias{lvhdr_norm}
\title{Normalize Model Parameters}
\usage{
lvhdr_norm(J, N, cov_len, px, tA, Theta, Gamma, tBeta, X)
}
\arguments{
\item{J}{Number of events or outcome variables.}

\item{N}{Number of subjects.}

\item{cov_len}{Total number of covariates (static and time-dependent).}

\item{px}{Number of static covariates.}

\item{tA}{Transposed A matrix.}

\item{Theta}{Theta matrix.}

\item{Gamma}{Gamma matrix.}

\item{tBeta}{Transposed Beta matrix representing covariate(static and time-dependent) effects.}

\item{X}{Matrix of static covariates.}
}
\value{
A list containing normalized versions of tA (tAstar), Theta (Thetastar), Gamma (Gammastar), and tBeta (tBetastar).
}
\description{
This function normalizes the model parameters (A, Theta, Gamma, and Beta) to satisfy identifiability conditions.
}
