% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{lvhml_est}
\alias{lvhml_est}
\title{Estimation Function for lvhml Model for binary response data.}
\usage{
lvhml_est(
  Y,
  R,
  X,
  Kset = 1:10,
  par = FALSE,
  n.cores = 1,
  Asymp = FALSE,
  Silent = FALSE,
  Z = array(0, dim = c(0, 0, 0)),
  full = FALSE,
  ext = FALSE,
  gamma_fix = FALSE,
  proj_const = 5
)
}
\arguments{
\item{Y}{A N x J x Tp array storing binary responses(1 or 0). Entries for unobserved data should be marked as NA.}

\item{R}{A N x Tp matrix representing the missing indicator (1 for observed, 0 for missing).}

\item{X}{A N x px matrix storing static covariates.}

\item{Kset}{A vector containing the candidate numbers of factors. The default value is 1:10.}

\item{par}{Logical. Indicates whether to use parallel computing. Default is FALSE.}

\item{n.cores}{The number of cores to use for parallel computing. Used in conjunction with par = TRUE. Default is 1.}

\item{Asymp}{Logical. Indicates whether to estimate the asymptotic variance. Default is FALSE.}

\item{Silent}{Logical. Option to suppress output showing progress. Default is FALSE.}

\item{Z}{A N x J x Tp array for time-dependent covariates. Defaults as empty array if not provided.}

\item{full}{Logical. Indicates whether full details are reported. Default is FALSE.}

\item{ext}{Logical. Indicates whether the extension allowing time-dependent loadings and coefficients on X is used. Default is FALSE.}

\item{gamma_fix}{Logical. Indicates whether the restriction gamma_{jt} = t gamma_{j} is imposed. Default is FALSE.}

\item{proj_const}{A scalar specifying the projection constant, also referred to as the constraint parameter in the manuscript. Default is 5.}
}
\value{
A list containing the results of the estimation, including:
        \item{Thetahat}{Estimated Theta}
        \item{Gammahat}{Estimated Gamma}
        \item{Betahat}{Estimated Beta}
        \item{Ahat}{Estimated A}
        \item{Khat}{Estimated K}
        \item{ICs}{Information criteria for values of K in Kset}
        \item{Sigma_E}{Estimated asymptotic variance matrices for sqrt(N)*betahat_j (if Asymp = TRUE)}
        \item{SE}{The standard errors of Betahat (if Asymp = TRUE)}
        \item{AsymVarhatstore}{The estimated variance matrices for sqrt(N)*u_j (if Asymp = TRUE)}
        \itme{all_output}{List containing all_output (if full = TRUE)}
}
\description{
Main function to implement the proposed estimator in a binary setting, as in simulation and data analysis. This includes choosing the number of factors (K) and calculating asymptotic variances if required. The function is capable of handling time-dependent covariates.
}
\examples{
# Generate example dataset
set.seed(123)
dataset <- lvhml_ex_data(J = 100, nset = "s", Kstar = 3,ext = FALSE, gamma_fix = FALSE, gap = FALSE)
attach(dataset)
out <- lvhml_est(Y, R, X, Kset = 1:10, par = FALSE, n.cores = 1, Asymp = FALSE, Silent = FALSE, Z = array(0, dim = c(0, 0, 0)),ext = FALSE,gamma_fix = FALSE)

}
