% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{lvhml_est_fit_K}
\alias{lvhml_est_fit_K}
\title{Internal Estimation Function for LVHML Model}
\usage{
lvhml_est_fit_K(
  Y,
  Ymatlog,
  R,
  X,
  Zmat,
  indices_R,
  lengths_R,
  Tp,
  K,
  px,
  pz,
  u_len,
  N,
  J,
  Tinit,
  Uinit,
  par,
  n.cores,
  Silent,
  ext,
  gamma_fix,
  proj_const
)
}
\arguments{
\item{Ymatlog}{Logical matrix derived from Y to reduce memory usage.}

\item{Zmat}{Matrix form time-dependent covariates Z.}

\item{indices_R}{Indices of observed entries in R.}

\item{lengths_R}{Number of observed time points for each subject.}

\item{Tp}{Number of time points.}

\item{K}{Number of factors to estimate.}

\item{px}{Number of static covariates.}

\item{pz}{Number of time-dependent covariates.}

\item{u_len}{Length. of the parameter u_j.}

\item{N}{Number of subjects.}

\item{J}{Number of events.}

\item{Tinit}{Initial values for Theta.}

\item{Uinit}{Initial values for U (Gamma,Beta, V, A).}
}
\value{
A list including:
  \item{success}{Convergence indicator (1 = success, 0 = failure).}
  \item{Thetahat}{Estimated Theta after final iteration.}
  \item{Uhat}{Estimated U after final iteration.}
  \item{Uit}{Array of U iterates (J x u_len x niter).}
  \item{Thetait}{Array of Theta iterates (N x K x niter).}
  \item{Improvementsobj}{Vector of objective improvements per iteration.}
  \item{niter}{Number of iterations executed.}
  \item{objstore}{Objective values per iteration.}
  \item{lastobj}{Objective at final iteration.}
  \item{total_proj}{Total number of projection operations performed.}
}
\description{
This function estimates A, Beta, and Theta for a given value of K. It's designed
to be used internally within the package and supports parallel computing, handling
of time-dependent covariates, and convergence checking.
}
\seealso{
\code{\link{lvhml_est}} for the description of \code{Y}, \code{R}, \code{X}, \code{par}, \code{n.cores}, \code{Silent}, \code{ext}, \code{gamma_fix} and \code{proj_const}.
}
\keyword{internal}
