% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_example_data.R
\name{lvhml_ex_data}
\alias{lvhml_ex_data}
\title{Function to generate example data.}
\usage{
lvhml_ex_data(
  J = 100,
  nset = "s",
  Kstar = 3,
  ext = FALSE,
  gamma_fix = FALSE,
  gap = FALSE
)
}
\arguments{
\item{J}{Number of event types.}

\item{nset}{sets the size of N relative to J, "s": N = 5J, "L":N = 10J}

\item{Kstar}{True latent dimension}

\item{ext}{Logical. Indicates whether the extension (section 2.4) allowing time-dependent loadings and coefficients on X is used. Default is FALSE.}

\item{gamma_fix}{Logical. Indicates whether constraint (section 2.5)  gamma_{jt} = t gamma_j  for j = 1,...,J is imposed. Default is FALSE.}

\item{gap}{Logical. Indicates whether the the setting (Section E.1) with sufficient eigengap for identification of latent factors is applied. Default is FALSE.}
}
\value{
A list that can be used as input for `lvhml_est` and true values of parameters for comparison.
        \item{Ustar}{A matrix storing the true values of U.}
        \item{Thetastar}{A matrix storing the true values of Theta.}
        \item{N,J,Tp}{Return values of N, J and Tp for referece.}
}
\description{
Function to generate example data.
}
\seealso{
\code{\link{lvhml_est}} for the description of the observed data Y, R and X.
}
