% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{lvhml_norm}
\alias{lvhml_norm}
\title{Normalize Model Parameters}
\usage{
lvhml_norm(
  J,
  N,
  px,
  tA,
  Theta,
  Gamma,
  tBeta,
  X,
  Tp,
  ext = FALSE,
  gamma_fix = FALSE
)
}
\arguments{
\item{J}{Number of events or outcome variables.}

\item{N}{Number of subjects.}

\item{px}{Number of static covariates.}

\item{tA}{Transposed loading matrix (K times J if ext = FALSE and Tp*K times J if ext = TRUE)}

\item{Theta}{(N x K) Matrix of latent factors.}

\item{Gamma}{(J x g_len) Matrix of intercepts. g_len = 1 if gamma_fix = TRUE and g_len = Tp otherwise.}

\item{tBeta}{Transposed regression coefficient matrix of X (px times J if ext = FALSE and Tp*px times J if ext = TRUE).}

\item{X}{(N x px) Matrix of static covariates.}

\item{ext}{Logical. Indicates whether the extension allowing time-dependent loadings and coefficients on X is used. Default is FALSE.}

\item{gamma_fix}{Logical. Indicates whether the restriction gamma_{jt} = t gamma_{j} is imposed. Default is FALSE.}
}
\value{
A list containing normalized versions of tA (tAstar), Theta (Thetastar), Gamma (Gammastar), and tBeta (tBetastar).

#' @return A list with normalized components:
  \item{tAstar}{Normalized transposed loadings.}
  \item{Thetastar}{Normalized latent factors}
  \item{Gammastar}{Normalized time-dependent intercept}
  \item{tBetastar}{Normalized transposed regression coefficients}
}
\description{
This function normalizes the model parameters (A, Theta, Gamma, and Beta) to satisfy identifiability conditions.
}
