% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{parglmm.func}
\alias{parglmm.func}
\title{Function to run glmm for each given j}
\usage{
parglmm.func(yris, indices_R, lengths_R, XZmat, N, P, K, Tp, ext, gamma_fix)
}
\arguments{
\item{yris}{(Tp x N) Matrix containing outcomes from Y associated with the jth event.
See \code{\link{lvhml_est}} for the meaning and dimensions of remaining arguments.}
}
\description{
Function to run glmm for each given j
}
