% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{run_svd}
\alias{run_svd}
\title{SVD-based low-rank reconstruction of the transformed responses}
\usage{
run_svd(mat, N, K, M_nrow, threshold)
}
\arguments{
\item{mat}{Numeric matrix (N times M_nrow) containing the transformed responses (stacked over time
if \code{ext = TRUE}).}

\item{N}{Number of subjects.}

\item{K}{Latent dimension.}

\item{M_nrow}{Number of columns of \code{mat} (Typically \code{J} if ext is FALSE and \code{J * Tp} otherwise).}

\item{threshold}{Thresold to be compared with singular values.}
}
\value{
Numeric matrix of the same dimension as \code{mat} containing the low–rank reconstruction.
}
\description{
Internal helper used in the SVD initialisation step
  to obtain a low–rank approximation of the transformed response
  matrix.
}
\keyword{internal}
