% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{Uest.func}
\alias{Uest.func}
\title{Update the Values of U (Gamma,Beta, V, A)}
\usage{
Uest.func(
  Y,
  R,
  U,
  Theta,
  X,
  Zmat,
  indices_R,
  lengths_R,
  Tp,
  K,
  px,
  pz,
  N,
  J,
  par,
  n.cores,
  ext,
  gamma_fix,
  proj_const
)
}
\arguments{
\item{U}{Current estimates of Gamma, A, and Beta parameters.}

\item{Theta}{Current estimates Theta.}
}
\value{
A matrix containing the updated value of U.
}
\description{
This function updates the values of U based on the current estimates.
It supports parallel processing for efficiency.
}
\seealso{
\code{lvhml_est_fit_K} for the description of other parameters.
}
\keyword{internal}
