% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_par.R
\name{Betaestglm.func}
\alias{Betaestglm.func}
\title{Beta Estimation for Initial Values}
\usage{
Betaestglm.func(
  Y,
  R,
  Gamma,
  A,
  Theta,
  XZmat,
  indices_R,
  lengths_R,
  Tp,
  K,
  cov_len,
  N,
  J,
  par,
  n.cores
)
}
\value{
A matrix containing the estimated Beta coefficients for each event.
}
\description{
This function estimates the initial values for Beta parameters using Generalized Linear Models (GLMs).
It supports parallel processing and is designed for internal package use.
}
\seealso{
\code{lvhdr_est_fit_K} for the description of parameters.
}
\keyword{internal}
