% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{Betaparglm.func}
\alias{Betaparglm.func}
\title{Beta (and V) Parameter Estimation with GLM}
\usage{
Betaparglm.func(
  yris,
  indices_R,
  lengths_R,
  gammaj,
  aj,
  Theta,
  X,
  Zmat,
  N,
  px,
  pz,
  K,
  Tp,
  ext,
  gamma_fix
)
}
\arguments{
\item{yris}{Transposed Y matrix for the jth event.}

\item{gammaj}{Gamma coefficients for the jth event.}

\item{aj}{A coefficients for the jth event.}
}
\value{
A vector of coefficients estimated by the GLM.
}
\description{
This internal function performs the estimation of Beta parameters for a single event using GLM.
It's utilized by \code{Betaestglm.func} for either sequential or parallel processing.
}
\seealso{
\code{lvhml_est_fit_K} for the description of other parameters.
}
\keyword{internal}
