% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_example_data.R
\name{lvhml_ex_data}
\alias{lvhml_ex_data}
\title{Function to generate example data.}
\usage{
lvhml_ex_data(J = 100, nset = "s", Kstar = 3)
}
\arguments{
\item{J}{Number of event types.}

\item{nset}{sets the size of N relative to J, "s": N = 5J, "L":N = 10J}

\item{Kstar}{True latent dimension}
}
\value{
A list structure that can be used as input for `lvhml_est` and true values of parameters for comparison.
        \item{Ustar}{A matrix storing the true values of U.}
        \item{Thetastar}{A matrix storing the true values of Theta.}
        \item{N,J,Tp}{Return values of N, J and Tp for referece.}
}
\description{
Function to generate example data.
}
\seealso{
\code{\link{lvhml_est}} for the description of the observed data Y, R and X.
}
