% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{SVDinit.func}
\alias{SVDinit.func}
\title{Compute Initial Values Using SVD-based Algorithm}
\usage{
SVDinit.func(
  Y,
  R,
  X,
  Zmat,
  indices_R,
  lengths_R,
  Tp,
  K,
  px,
  pz,
  N,
  J,
  par,
  n.cores,
  ext,
  gamma_fix
)
}
\value{
A list containing initial values for U (Gamma,Beta, V, A) and T (Theta).
}
\description{
This function calculates initial values for parameters using a Singular Value Decomposition (SVD)-based algorithm.
It's intended for internal use and supports parallel computing for efficient operation.
}
\seealso{
\code{lvhml_est_fit_K} for the description of parameters.
}
\keyword{internal}
