% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{Thetaest.func}
\alias{Thetaest.func}
\title{Update Theta Parameters}
\usage{
Thetaest.func(
  Y,
  R,
  Theta,
  U,
  X,
  Z,
  Tp,
  K,
  px,
  pz,
  N,
  J,
  par,
  n.cores,
  ext,
  gamma_fix,
  proj_const
)
}
\arguments{
\item{Theta}{Current (N x K) matrix of latent factors.}

\item{U}{Current (J x u_len) matrix of event-specific parameters,
where each row contains the stacked vector \code{(Gamma_j, Beta_j, V_j, A_j)}.}
}
\value{
A matrix of updated Theta parameters for each subject.
}
\description{
This function updates the Theta parameters based on the current estimates of U, Theta, and other model parameters.
It supports parallel processing for efficient computation.
}
\seealso{
\code{lvhml_est_fit_K} for the description of other parameters.
}
\keyword{internal}
