% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_par.R
\name{Thetaest.func}
\alias{Thetaest.func}
\title{Update Theta Parameters}
\usage{
Thetaest.func(Y, R, Theta, U, XZ, Tp, K, cov_len, N, J, par, n.cores)
}
\arguments{
\item{Theta}{Current estimates Theta.}

\item{U}{Current estimates of Gamma, A, and Beta parameters.}
}
\value{
A matrix of updated Theta parameters for each subject.
}
\description{
This function updates the Theta parameters based on the current estimates of U, Theta, and other model parameters.
It supports parallel processing for efficient computation.
}
\seealso{
\code{lvhml_est_fit_K} for the description of other parameters.
}
\keyword{internal}
