% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{Thetapar.func}
\alias{Thetapar.func}
\title{Parallel Function for Theta Parameter Update}
\usage{
Thetapar.func(
  yrjs,
  Ri,
  thetai,
  A,
  Gamma,
  Beta,
  V,
  X,
  Z,
  K,
  Tp,
  px,
  pz,
  i,
  J,
  ext,
  gamma_fix,
  proj_const
)
}
\arguments{
\item{yrjs}{Transposed Y matrix for the ith subject.}

\item{Ri}{Missing indicator vector for the ith subject.}

\item{thetai}{Current Theta estimate for the ith subject.}

\item{A}{Current estimate of A.}

\item{Gamma}{Current estimate of Gamma.}

\item{Beta}{Current estimate of Beta.}

\item{i}{Subject index.

#'@seealso \code{lvhml_est_fit_K} for the description of other parameters.}
}
\value{
A vector of updated Theta parameters for the subject.
}
\description{
Computes the update for Theta parameters for a single subject using current model estimates.
}
\keyword{internal}
