% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_par.R
\name{Upar.func}
\alias{Upar.func}
\title{Parallel Function for U Parameter Update}
\usage{
Upar.func(
  yris,
  indices_R,
  lengths_R,
  uj,
  Theta,
  X,
  Zmat,
  N,
  px,
  pz,
  K,
  Tp,
  ext,
  gamma_fix,
  proj_const
)
}
\arguments{
\item{yris}{Transposed Y matrix for the jth event.}

\item{uj}{Current estimates of the jth row of U for the event.}

\item{Theta}{Current estimates Theta.}
}
\value{
A vector of updated U parameters for the event.
}
\description{
Performs the update of U parameters for a single event. It calculates gradients and
Hessians to find the optimal update direction and magnitude for U parameters.
}
\seealso{
\code{lvhml_est_fit_K} for the description of other parameters.
}
\keyword{internal}
